# Conflicts:
#	README.md
#	README_en.md
#	VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Editor/BuildHelper.cs
#	VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Utils/StreamingAssetFileNames.cs

+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Editor/BuildHelper.cs
@@ -152,10 +152,11 @@ namespace Baku.VMagicMirror
                 return;
             }
 
-            //VRMLoaderUI以外のディレクトリは削除
+            // フォルダ: VRMLoaderUI, MediaPipeTrackerだけがruntimeで必要
             foreach (var dir in Directory.GetDirectories(streamingAssetDir))
             {
-                if (Path.GetFileName(dir) != StreamingAssetFileNames.LoaderUiFolder)
+                if (Path.GetFileName(dir) != StreamingAssetFileNames.LoaderUiFolder &&
+                    Path.GetFileName(dir) != StreamingAssetFileNames.MediaPipeTrackerFolder)
                 {
                     Directory.Delete(dir, true);
                 }
